requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
packaging==23.1
pluggy==1.0.0
pydub==0.25.1
pytest==7.3.2
pytest-asyncio==0.21.1
pytest-mock==3.10.0
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.3.0
tomli==2.0.1
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@a2534f7b888416e872d5afd1862eb3e30fc69fc7
